Update AssetBundleTree.cs
While creating my own AssetBundlePackagerBrowserIntergration I noticed hitting the refresh button in the top left corner of the asset bundle browser caused the GetAllAssetBundleNames method to be called twice. I'm doing some expensive operations in that method so it's pretty noticeable when it takes twice as long.

Looks like the AssetBundleInspectorTab was making a call to m_BundleTreeView.Reload(); which eventual calls Refresh and the AssetBundleTree class was making a call to AssetBundleModel.Model.Refresh(); causing the Refresh method to be invoked twice.

I chose to remove the AssetBundleModel.Model.Refresh(); line which resolved the issue for me.